requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
pymongo==4.5.0
bcrypt==4.1.3
python-jose==3.5.0
python-dotenv==1.2.1
python-multipart==0.0.21
passlib==1.7.4
resend==2.19.0
dnspython==2.8.0
